refactor: change logic to formOrder hook
